requirements.txt
agate==1.6.1
agate-dbf==0.2.1
agate-excel==0.2.3
agate-sql==0.5.4
alabaster==0.7.12
anaconda-client==1.7.2
anaconda-navigator==1.9.7
anaconda-project==0.8.2
appdirs==1.4.3
asn1crypto==0.24.0
astroid==2.2.5
astropy==3.1.2
atomicwrites==1.3.0
attrs==19.1.0
autograd==1.2
Babel==2.6.0
backcall==0.1.0
backoff==1.8.0
backports.os==0.1.1
backports.shutil-get-terminal-size==1.0.0
basilica==0.2.6
beautifulsoup4==4.7.1
bitarray==0.9.2
bkcharts==0.2
black==19.3b0
blaze==0.11.3
bleach==3.1.0
blinker==1.4
bokeh==1.1.0
boto==2.49.0
boto3==1.9.134
botocore==1.12.134
Bottleneck==1.2.1
bz2file==0.98
catboost==0.14.2
category-encoders==2.0.0
cchardet==1.1.3
certifi==2019.3.9
cffi==1.12.3
chardet==3.0.4
Click==7.0
cloudpickle==1.0.0
clyent==1.2.2
colorama==0.4.1
conda==4.6.14
conda-build==3.17.8
conda-package-handling==0+unknown
conda-verify==3.3.0
configparser==3.7.4
contextlib2==0.5.5
coverage==4.5.3
cryptography==2.6.1
csvkit==1.0.4
cycler==0.10.0
Cython==0.29.7
cytoolz==0.9.0.1
dask==1.2.2
datadotworld==1.6.0
datapackage==0.8.9
datashape==0.5.4
dbfread==2.0.7
decorator==4.4.0
defusedxml==0.6.0
distributed==1.28.1
docutils==0.14
dwcontents==1.0.0b5
eli5==0.8.2
entrypoints==0.3
et-xmlfile==1.0.1
ezodf==0.3.2
fastcache==1.1.0
filelock==3.0.10
flake8==3.7.7
flake8-import-order==0.18.1
flake8-polyfill==1.0.2
Flask==1.0.2
Flask-Cors==3.0.7
Flask-SQLAlchemy==2.3.2
future==0.17.1
gensim==3.4.0
gevent==1.4.0
glob2==0.6
gmpy2==2.0.8
graphviz==0.10.1
greenlet==0.4.15
gunicorn==19.9.0
h5py==2.9.0
heapdict==1.0.0
html5lib==1.0.1
idna==2.8
ijson==2.3
imageio==2.5.0
imagesize==1.1.0
importlib-metadata==0.0.0
ipykernel==5.1.0
ipython==7.5.0
ipython-genutils==0.2.0
ipywidgets==7.4.2
isodate==0.6.0
isort==4.3.19
itsdangerous==1.1.0
jdcal==1.4.1
jedi==0.13.3
jeepney==0.4
Jinja2==2.10.1
jmespath==0.9.4
joblib==0.13.2
jsonlines==1.2.0
jsonschema==3.0.1
jsontableschema==0.10.1
jupyter==1.0.0
jupyter-client==5.2.4
jupyter-console==6.0.0
jupyter-core==4.4.0
jupyterlab==0.35.5
jupyterlab-launcher==0.13.1
jupyterlab-server==0.2.0
kaggle==1.5.3
keyring==18.0.0
kiwisolver==1.1.0
lambdata-alex-witt==0.0.2
lambdata-higgins2718==0.0.1
lambdata-standroidbeta==0.2.0
lazy-object-proxy==1.4.1
leather==0.3.3
libarchive-c==2.8
lief==0.9.0
lifelines==0.21.2
lightgbm==2.2.1
linear-tsv==1.1.0
llvmlite==0.28.0
locket==0.2.0
lxml==4.3.3
MarkupSafe==1.1.1
matplotlib==3.1.0
mccabe==0.6.1
mistune==0.8.4
mkl-fft==1.0.12
mkl-random==1.0.2
mlxtend==0.13.0
mock==3.0.5
more-itertools==7.0.0
mpmath==1.1.0
msgpack==0.6.1
multipledispatch==0.6.0
navigator-updater==0.2.1
-e git+https://github.com/jupyter/nbconvert.git@4d8efc0b84cc0a29509f301ef260b07eb9ec668e#egg=nbconvert
nbformat==4.4.0
networkx==2.3
nltk==3.4.1
nose==1.3.7
notebook==5.7.8
numba==0.43.1
numexpr==2.6.9
numpy==1.16.3
numpydoc==0.9.1
oauthlib==3.0.1
odo==0.5.1
olefile==0.46
openpyxl==2.6.2
packaging==19.0
pandas==0.23.4
pandocfilters==1.4.2
parsedatetime==2.4
parso==0.4.0
partd==0.3.10
path.py==12.0.1
pathlib2==2.3.3
patsy==0.5.1
PDPbox==0.2.0
pep8==1.7.1
pexpect==4.7.0
pickleshare==0.7.5
Pillow==6.0.0
pipenv==2018.11.26
pkginfo==1.5.0.1
pluggy==0.11.0
ply==3.11
prometheus-client==0.6.0
prompt-toolkit==2.0.9
psutil==5.6.2
psycopg2==2.7.6.1
psycopg2-binary==2.8.2
ptyprocess==0.6.0
py==1.8.0
pycodestyle==2.5.0
pycosat==0.6.3
pycparser==2.19
pycrypto==2.6.1
pycurl==7.43.0.2
pyflakes==2.1.1
Pygments==2.4.2
PyJWT==1.7.1
pylint==2.3.1
pyodbc==4.0.26
pyOpenSSL==19.0.0
pyparsing==2.4.0
PyQt5==5.12.2
PyQt5-sip==4.19.17
PyQtWebEngine==5.12.1
pyrsistent==0.14.11
PySocks==1.7.0
pytest==4.5.0
pytest-arraydiff==0.3
pytest-astropy==0.5.0
pytest-cov==2.7.1
pytest-doctestplus==0.3.0
pytest-openfiles==0.3.2
pytest-remotedata==0.3.1
python-crfsuite==0.9.6
python-dateutil==2.8.0
python-decouple==3.1
python-dotenv==0.10.3
python-slugify==3.0.2
pytimeparse==1.1.8
pytz==2019.1
PyWavelets==1.0.3
PyYAML==5.1
pyzmq==18.0.0
QtAwesome==0.5.7
qtconsole==4.3.1
QtPy==1.7.1
readme-renderer==24.0
reportlab==3.5.19
requests==2.22.0
requests-oauthlib==1.2.0
requests-toolbelt==0.9.1
rfc3986==0.4.1
rope==0.14.0
ruamel-yaml==0.15.46
s3transfer==0.2.0
scikit-image==0.15.0
scikit-learn==0.21.1
scipy==1.3.0
seaborn==0.9.0
SecretStorage==3.1.1
Send2Trash==1.5.0
shap==0.28.5
simplegeneric==0.8.1
simplejson==3.16.0
singledispatch==3.4.0.3
six==1.12.0
sklearn==0.0
smart-open==1.8.2
snowballstemmer==1.2.1
sortedcollections==1.1.2
sortedcontainers==2.1.0
soupsieve==1.8
Sphinx==2.0.1
sphinxcontrib-applehelp==1.0.1
sphinxcontrib-devhelp==1.0.1
sphinxcontrib-htmlhelp==1.0.2
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.2
sphinxcontrib-serializinghtml==1.1.3
sphinxcontrib-websupport==1.1.0
spyder==3.3.4
spyder-kernels==0.4.4
SQLAlchemy==1.3.3
stability-selection==0.0.1
statsmodels==0.9.0
sympy==1.4
tables==3.5.1
tabulate==0.8.3
tabulator==1.4.1
tblib==1.4.0
terminado==0.8.2
testpath==0.4.2
text-unidecode==1.2
toml==0.10.0
toolz==0.9.0
tornado==6.0.2
tqdm==4.32.1
traitlets==4.3.2
tweepy==3.7.0
twine==1.13.0
twython==3.7.0
typed-ast==1.3.4
typing==3.6.6
unicodecsv==0.14.1
Unidecode==1.0.23
urllib3==1.25.3
virtualenv==16.6.0
virtualenv-clone==0.5.3
wcwidth==0.1.7
webencodings==0.5.1
Werkzeug==0.15.2
Whoosh==2.7.4
widgetsnbextension==3.4.2
wrapt==1.11.1
wurlitzer==1.0.2
xlrd==1.2.0
XlsxWriter==1.1.8
xlwt==1.3.0
zict==0.1.4
zipp==0.3.3
